models on models mostly amount fixes

+++ internal/models/ticket_purchase.go
@@ -8,8 +8,8 @@ type TicketPurchase struct {
 	ID          string `gorm:"type:uuid;default:gen_random_uuid();primaryKey"`
 	UserID      string    `gorm:"type:uuid"`
 	TransactionHistoryID string   `gorm:"type:uuid"`
-	TotalAmount float64
-	UnitPrice   float64
+	TotalAmount int64
+	UnitPrice   int64
 	Quantity    int
 	PurchasedAt time.Time `gorm:"default:current_timestamp"`
 	Currency    string    `gorm:"size:3;default:'NGN'"`
